Extract default webview configuration.
