+++ Core/Core/Courses/K5/ViewModel/K5SubjectViewModel.swift
@@ -31,7 +31,7 @@ public class K5SubjectViewModel: ObservableObject {
     public private(set) lazy var reloadWebView: AnyPublisher<Void, Never> = makeWebViewReloadTrigger()
 
     /** The webview configuration to be used. In case of masquerading we can't use the default configuration because it will contain cookies with the original user's permissions. */
-    public var config: WKWebViewConfiguration? { masqueradedSession.config }
+    public var config: WKWebViewConfiguration { masqueradedSession.config }
 
     private let env = AppEnvironment.shared
     private let context: Context
